address moderator changes
The <owners> field is for the package owners and should be your name / Chocolatey username.

Can you use a static CDN for your <iconUrl> - githack.com and jsdelivr.com are popular.

Can you provide a proper <description> and <summary>?

Can you remove the code that checks whether the software is already installed. Chocolatey will only install a package if it's not already installed, unless the --force switch is provided.